Rename Annotation*Field classes
They are really about foreign fields
