+++ lightly_studio/src/lightly_studio/core/dataset_query/classification_expression.py
@@ -7,8 +7,8 @@ from dataclasses import dataclass
 from sqlalchemy import ColumnElement, and_
 from sqlmodel import col
 
-from lightly_studio.core.dataset_query.annotation_expression import AnnotationComparableField
 from lightly_studio.core.dataset_query.boolean_expression import AND
+from lightly_studio.core.dataset_query.foreign_field import ForeignComparableField
 from lightly_studio.core.dataset_query.match_expression import MatchExpression
 from lightly_studio.models.annotation.annotation_base import (
     AnnotationBaseTable,
@@ -21,7 +21,7 @@ from lightly_studio.models.sample import SampleTable
 class ClassificationField:
     """Providing access to predefined classification fields for queries."""
 
-    label = AnnotationComparableField(
+    label = ForeignComparableField(
         column=col(AnnotationLabelTable.annotation_label_name),
         relationship=AnnotationBaseTable.annotation_label,
     )
